src/main/java/com/saulo/webstore/controllers/ContaController.java
package com.saulo.webstore.controllers;

import com.saulo.webstore.models.Categoria;
import com.saulo.webstore.models.Conta;
import com.saulo.webstore.services.CategoriaService;
import com.saulo.webstore.services.ContaService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RestController;

import java.util.ArrayList;
import java.util.List;

@RestController
@RequestMapping(value = "/contas")
public class ContaController {

    @Autowired
    private ContaService contaService;

    @RequestMapping(value="/{id}", method=RequestMethod.GET)
    public ResponseEntity<Conta> findById(@PathVariable Integer id) {
        Conta obj = contaService.findById(id);
        return ResponseEntity.ok().body(obj);
    }

}
